requirements.txt
motor==3.4.0
python-dotenv==1.0.1
aiogram==3.6.0